bookstore/requirements.txt
amqp==2.2.2
asn1crypto==0.24.0
billiard==3.5.0.3
celery==4.1.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==1.11
django-cors-headers==2.4.0
django-redis==4.8.0
django-tinymce==2.6.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
jieba==0.39
kombu==4.1.0
olefile==0.44
Pillow==4.3.0
pycparser==2.19
pycryptodome==3.4.7
pycryptodomex==3.6.6
PyJWT==1.6.4
PyMySQL==0.9.2
python-alipay-sdk==1.7.0
pytz==2017.2
redis==2.10.6
requests==2.19.1
six==1.11.0
urllib3==1.23
vine==1.1.4
Whoosh==2.7.4
django-haystack==2.8.1
